pyproject.toml
[tool.poetry]
name = "nginx-http-stats"
version = "0.1.0"
description = ""
authors = ["Sven Ulland <sven.ulland+github@gmail.com>"]
readme = "README.md"
packages = [{include = "nginx_http_stats"}]

[tool.poetry.scripts]
nginx-http-stats = "nginx_http_stats.cli:cli"

[tool.poetry.dependencies]
python = "^3.8"
python-json-logger = "^2.0.4"
PyYAML = "^6.0"

[tool.poetry.group.dev.dependencies]
black = "^22.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
